+++ R/content.R
@@ -142,7 +142,7 @@ ContentImage <- new_class(
 #' @param detail Not currently used.
 ContentImageRemote <- new_class(
   "ContentImageRemote",
-  parent = Content,
+  parent = ContentImage,
   properties = list(
     url = prop_string(),
     detail = prop_string(default = "")
@@ -164,7 +164,7 @@ method(contents_markdown, ContentImageRemote) <- function(content) {
 #' @param data Base64 encoded image data.
 ContentImageInline <- new_class(
   "ContentImageInline",
-  parent = Content,
+  parent = ContentImage,
   properties = list(
     type = prop_string(),
     data = prop_string(allow_null = TRUE)
@@ -318,7 +318,18 @@ tool_string <- function(x) {
   } else if (is.character(x@value)) {
     paste(x@value, collapse = "\n")
   } else {
-    jsonlite::toJSON(x@value, auto_unbox = TRUE)
+    withCallingHandlers(
+      jsonlite::toJSON(x@value, auto_unbox = TRUE),
+      error = function(err) {
+        cli::cli_abort(
+          c(
+            "Could not convert tool result from {.obj_type_friendly {x@value}} to JSON.",
+            "i" = "If you are the tool author, update the tool to convert the result to a string or JSON."
+          ),
+          parent = err
+        )
+      }
+    )
   }
 }