+++ R/provider-aws.R
@@ -370,7 +370,11 @@ method(as_json, list(ProviderAWSBedrock, Turn)) <- function(provider, x, ...) {
     # bedrock passes system prompt as separate arg
     NULL
   } else if (is_user_turn(x) || is_assistant_turn(x)) {
-    list(role = x@role, content = as_json(provider, x@contents, ...))
+    x <- turn_contents_expand(x)
+    list(
+      role = x@role,
+      content = as_json(provider, x@contents, ...)
+    )
   } else {
     cli::cli_abort("Unknown role {x@role}", .internal = TRUE)
   }