+++ R/provider-openai.R
@@ -330,6 +330,8 @@ method(as_json, list(ProviderOpenAI, Turn)) <- function(
   x,
   ...
 ) {
+  x <- turn_contents_expand(x)
+
   # While the user turn can contain multiple contents, the assistant turn
   # can't. Fortunately, we can send multiple user turns with out issue.
   as_json(provider, x@contents, ..., role = x@role)