+++ R/provider-snowflake.R
@@ -286,6 +286,7 @@ method(as_json, list(ProviderSnowflakeCortex, Turn)) <- function(
   } else {
     cli::cli_abort("Unsupported content type: {.cls {class(x@contents[[1]])}}.")
   }
+  x <- turn_contents_expand(x)
   list(
     role = x@role,
     content_list = as_json(provider, x@contents, ...)